api/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
awscli==1.32.72
blinker==1.7.0
boto3==1.34.72
botocore==1.34.72
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.22.1
idna==3.6
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
jmespath==1.0.1
keras==3.1.0
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.7
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors==0.4.2
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
split-folders==0.5.1
SQLAlchemy==2.0.29
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.35
urllib3==2.2.1
voluptuous==0.14.2
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yolo==0.3.2
